refactor(window): narrow chrome replacement fix
