+++ src/electron/nativeWindowChrome.js
@@ -1,15 +1,6 @@
 'use strict';
 
 const NATIVE_TITLEBAR_HEIGHT = 30;
-const DEFAULT_SYSTEM_WINDOW_CONTROLS = false;
-
-function windowChromeState(settings = {}, options = {}) {
-  return {
-    systemWindowControls: settings.systemWindowControls === true,
-    collapsedFloatingBubble: options.collapsedFloatingBubble === true,
-    trayOnly: settings.trayMode === true,
-  };
-}
 
 function windowChromeTransition(previousSettings = {}, nextSettings = {}, options = {}) {
   const trayModeChanged = nextSettings.trayMode !== previousSettings.trayMode;
@@ -24,34 +15,6 @@ function windowChromeTransition(previousSettings = {}, nextSettings = {}, option
       ...(trayModeChanged && !nextSettings.trayMode ? { focus: true } : {}),
     },
   };
-}
-
-function windowChromeRequiresReplacement(platform, chrome = {}) {
-  return platform === 'win32' || chrome.systemWindowControls === true;
-}
-
-function observeMainFrameLoad(webContents, onSettled) {
-  let settled = false;
-  const cleanup = () => {
-    webContents.removeListener('did-finish-load', onFinished);
-    webContents.removeListener('did-fail-load', onFailed);
-  };
-  const settle = (result) => {
-    if (settled) return false;
-    settled = true;
-    cleanup();
-    onSettled(result);
-    return true;
-  };
-  const onFinished = () => settle({ ok: true });
-  const onFailed = (_event, errorCode, errorDescription, _url, isMainFrame) => {
-    if (!isMainFrame || errorCode === -3) return;
-    settle({ ok: false, errorCode, errorDescription });
-  };
-
-  webContents.once('did-finish-load', onFinished);
-  webContents.on('did-fail-load', onFailed);
-  return { settle, dispose: cleanup };
 }
 
 function nativeWindowChromeOptions(platform = process.platform, options = {}) {
@@ -74,11 +37,7 @@ function nativeWindowChromeOptions(platform = process.platform, options = {}) {
 }
 
 module.exports = {
-  DEFAULT_SYSTEM_WINDOW_CONTROLS,
   NATIVE_TITLEBAR_HEIGHT,
   nativeWindowChromeOptions,
-  observeMainFrameLoad,
-  windowChromeRequiresReplacement,
-  windowChromeState,
   windowChromeTransition,
 };
